requirements.txt
praw==6.5.1
python-dotenv==0.13.0